Se agrega el test de varias sumas y restas pasan el test

+++ src/main/java/co/com/sofka/app/calculator/BasicCalculator.java
@@ -10,7 +10,7 @@ public class BasicCalculator {
         logger.info( "Summing {} + {}", number1, number2 );
         return number1 + number2;
     }
-    public Long resta(Long number1, Long number2) {
+    public Double resta(Double number1, Double number2) {
         logger.info( "Summing {} - {}", number1, number2 );
         return number1 - number2;
     }
